damn
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@245 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/nrrd/write.c
@@ -148,7 +148,7 @@ _nrrdWriteDataRaw(Nrrd *nrrd, nrrdIO *io) {
   }
   else {
     if (AIR_DIO && nrrdFormatNRRD == io->format) {
-      fprintf(stderr, "with fwrite()%d ... ", dio); fflush(stderr);
+      fprintf(stderr, "with fwrite()%d ... ", (int)dio); fflush(stderr);
     }
     ret = fwrite(nrrd->data, nrrdElementSize(nrrd), nrrd->num, io->dataFile);
     if (ret != nrrd->num) {
